Use ASDF-managed python in LSP-pyright
The builtin lookup function in `lsp-pyright` isn't aware of ASDF shims,
so when it passes a `python` binary to the pyright server, the server
executes the shim somewhere in `~/.emacs.d`, where the shim points to
the wrong python version.

This helper short-circuits the python lookup and replaces it with the
result of `asdf which python` (which gives an unshimmed path) if
applicable.