Add type & theme args to ob-plantuml blocks
This allows specifying a diagram type and them via header args in
plantuml source blocks rather than requiring those be specified via
directives in the source block itself.

The built-in function to create the complete body of a plantuml block in
ob-plantuml looks for a `@start` keyword at the beginning of a block,
and if none is present it unilaterally wraps the block in
`@startuml`/`@enduml`.  However, when we _do_ specify a diagram via
an inline `@start<type>`, this causes variable (and now theme)
directives to be ignored from the header args, because they'd fall
outside this scope of the diagram start/end tags.

By adding this type parameter we can neatly solve this problem with
directives from header args, which also conveniently allows us to (for
example) specify a common theme as file or node-level header args to be
used across multiple diagrams.